backend/requirements.txt
# aiohttp==3.9.5
# aiosignal==1.3.1
# annotated-types==0.7.0
# anyio==4.4.0
# attrs==23.2.0
# certifi==2024.7.4
# charset-normalizer==3.3.2
# click==8.1.7
# colorama==0.4.6
# distro==1.9.0
# dnspython==2.6.1
# email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
# frozenlist==1.4.1
# greenlet==3.0.3
# groq==0.9.0
# h11==0.14.0
# httpcore==1.0.5
# httptools==0.6.1
# httpx==0.27.0
# idna==3.7
# Jinja2==3.1.4
# jsonpatch==1.33
# jsonpointer==3.0.0
langchain==0.2.7
# langchain-core==0.2.13
langchain-groq==0.1.6
# langchain-text-splitters==0.2.2
# langsmith==0.1.85
# markdown-it-py==3.0.0
# MarkupSafe==2.1.5
# mdurl==0.1.2
# multidict==6.0.5
# numpy==1.26.4
# orjson==3.10.6
# packaging==24.1
# pydantic==2.8.2
# pydantic_core==2.20.1
# Pygments==2.18.0
# python-dotenv==1.0.1
# python-multipart==0.0.9
# PyYAML==6.0.1
# requests==2.32.3
# rich==13.7.1
# shellingham==1.5.4
# sniffio==1.3.1
# # SQLAlchemy==2.0.31
# starlette==0.37.2
# tenacity==8.5.0
# typer==0.12.3
# typing_extensions==4.12.2
# ujson==5.10.0
# urllib3==2.2.2
# uvicorn==0.30.1
# watchfiles==0.22.0
# websockets==12.0
# yarl==1.9.4
gTTS==2.5.2
better-profanity==0.7.0
# langchain
# langchain_groq
langchain_anthropic
# python-dotenv
langchain-openai
langchain-google-genai
# langchain_nvidia_ai_endpoints
Pillow==10.4.0

# # END